fix(cp/miniapp/common): P3 impl 层语义补齐——8 项逐项对齐 Java 4.8.6
- cp 智能机器人（#4100）：补齐 parseEncryptedCallbackMessage + replyMessage + crypt util
- cp agent_id（#4079）：i32 → i64 长整型应用 ID
- miniapp 手机号 openid（#4078）：新增 get_phone_number_with_openid default 方法
- miniapp getUserEncryptKey（4.8.5 修复）：签名算法对齐——session_key 直接 UTF-8 字节
- 已对齐确认：cp 待办 API / NFC scheme / JSON 验签；open 商家客服归 channel 模块
- cp 521 / miniapp 329 / open 190 tests 全绿

+++ crates/wx-rust-cp/src/bean/wx_cp_agent.rs
@@ -12,7 +12,7 @@ pub struct WxCpAgent {
     #[serde(rename = "errmsg", default)]
     pub err_msg: String,
     #[serde(rename = "agentid", default)]
-    pub agent_id: i32,
+    pub agent_id: i64,
     #[serde(rename = "name", default)]
     pub name: String,
     #[serde(rename = "square_logo_url", default)]
